UnifiedEmail: Add native heads up support per account (2/2)
Add native heads up notification support. User can set on every account
if he or she want this account as heads up notification or not.

This patch adds the settings for it as well the complete handling. Due
of a half broken NotificationCompat.Builder class we switch here to the
Notification.Builder class to get access to the methods we need. This is no
problem here due that we do not need compatibility on lower API's

Change-Id: I32cd1b914777dd3a491ab20fe681d347a9d81e3c
Signed-off-by: Josue Rivera <prbassplayer@gmail.com>